pyproject.toml
[tool.poetry]
name = "my-salary"
version = "0.1.0"
description = ""
authors = ["vivera83 <vivera83@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.18"
uvicorn = "^0.22.0"
pydantic = "^2.0.3"
pydantic-settings = "^2.0.2"
fastapi = "^0.100.0"
loguru = "^0.7.0"
itsdangerous = "^2.1.2"
python-jose = "^3.3.0"
email-validator = "^2.0.0.post2"
fastapi-jwt = "^0.1.12"
asyncpg = "^0.28.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
